+++ Assets/scripts/engine/UpgradeableObjects/BombCooldownUpgrader.cs
@@ -8,7 +8,7 @@ public class BombCooldownUpgrader : UpgradeableObject
 
   public override void RecalculateUpgradeCost()
   {
-    upgrade_cost = (int)((Mathf.Pow((float)(value + 1), 1.3f) * 750));
+    upgrade_cost = (ulong)((Mathf.Pow((float)(value + 1), 1.3f) * 750));
   }
 
   public override void Upgrade()
