+++ Assets/scripts/engine/UpgradeableObjects/BumperUpgrader.cs
@@ -5,7 +5,7 @@ public class BumperUpgrader : ScoringObjectUpgrade
 {
 	public override void RecalculateUpgradeCost()
 	{
-    upgrade_cost = (int)((Mathf.Pow((float)(value + 1), 1.4f) * 150));
+    upgrade_cost = (ulong)((Mathf.Pow((float)(value + 1), 1.4f) * 150));
 	}
 
   public override void RecalculatePointValue()
