+++ Assets/scripts/engine/UpgradeableObjects/CoinUpgrader.cs
@@ -8,7 +8,7 @@ public class CoinUpgrader : ScoringObjectUpgrade
 
 	public override void RecalculateUpgradeCost()
 	{
-		upgrade_cost = (int)((Mathf.Pow((float)(value + 1), 1.3f) * 50)); 
+    upgrade_cost = (ulong)((Mathf.Pow((float)(value + 1), 1.3f) * 50)); 
 	}
 
   public override void RecalculatePointValue()
