+++ Assets/scripts/engine/UpgradeableObjects/MaxBombsUpgrader.cs
@@ -7,7 +7,7 @@ public class MaxBombsUpgrader : UpgradeableObject
   {
     int starting_index_adjuster = default_value_on_reset - 1;
 
-    upgrade_cost = (int)((Mathf.Pow((float)(value - starting_index_adjuster), 1.4f) * 1000));
+    upgrade_cost = (ulong)((Mathf.Pow((float)(value - starting_index_adjuster), 1.4f) * 1000));
   }
 
   public override void Upgrade()
