+++ Assets/scripts/engine/UpgradeableObjects/PegUpgrader.cs
@@ -5,7 +5,7 @@ public class PegUpgrader : ScoringObjectUpgrade
 {
 	public override void RecalculateUpgradeCost()
 	{
-		upgrade_cost = (int)((Mathf.Pow((float)(value + 1), 1.4f) * 200));
+    upgrade_cost = (ulong)((Mathf.Pow((float)(value + 1), 1.4f) * 200));
 	}
 
   public override void RecalculatePointValue()
