+++ Assets/scripts/engine/UpgradeableObjects/UpgradeableObject.cs
@@ -1,17 +1,17 @@
 ﻿using UnityEngine;
 using System.Collections;
 
-abstract public class UpgradeableObject : SavedStat 
+abstract public class UpgradeableObject : SavedStat
 {
   [SerializeField]
   protected int max_upgrades = 0;
 
   protected bool is_plinkagon_upgrade = false;
 
-	protected int upgrade_cost;
+  protected ulong upgrade_cost;
 
-	abstract public void Upgrade();
-	abstract public void RecalculateUpgradeCost();
+  abstract public void Upgrade();
+  abstract public void RecalculateUpgradeCost();
 
   new protected virtual void OnEnable()
   {
@@ -28,28 +28,28 @@ abstract public class UpgradeableObject : SavedStat
     RecalculateUpgradeCost();
   }
 
-	void OnDestroy()
-	{
-		Save ();
-	}
+  void OnDestroy()
+  {
+    Save();
+  }
 
-	public int GetUpgrades()
-	{
-		return GetValue();
-	}
+  public int GetUpgrades()
+  {
+    return GetValue();
+  }
 
-	public int GetUpgradeCost()
-	{
-		return upgrade_cost;
-	}
+  public ulong GetUpgradeCost()
+  {
+    return upgrade_cost;
+  }
 
-	public virtual bool PlayerHasEnoughCurrency()
-	{
+  public virtual bool PlayerHasEnoughCurrency()
+  {
     if (!is_plinkagon_upgrade)
       return (MoneyTracker.Instance.GetCurrentMoney() >= upgrade_cost);
     else
       return (MoneyTracker.Instance.GetCurrentPlinkagonPoints() >= upgrade_cost);
-	}
+  }
 
   public int GetUpgradesOnReset()
   {
